+++ src/main/java/com/plivo/api/models/maskingsession/MaskingSessionUpdater.java
@@ -5,7 +5,8 @@ import com.plivo.api.models.base.VoiceUpdater;
 import retrofit2.Call;
 
 public class MaskingSessionUpdater extends VoiceUpdater<MaskingSessionUpdateResponse> {
-
+  private String firstParty;
+  private String secondParty;
   private int sessionExpiry;
   private int callTimeLimit;
   private int ringTimeout;
@@ -19,7 +20,7 @@ public class MaskingSessionUpdater extends VoiceUpdater<MaskingSessionUpdateResp
   private String recordFileFormat;
   private Boolean geomatch = null;
   private boolean record;
-
+  private boolean createSessionWithSingleParty;
   public MaskingSessionUpdater(String id) {
     super(id);
   }
@@ -36,6 +37,20 @@ public class MaskingSessionUpdater extends VoiceUpdater<MaskingSessionUpdateResp
     return client().getVoiceFallback2Service().maskingSessionUpdate(client().getAuthId(), id, this);
   }
 
+  public String firstParty() {
+    return this.firstParty;
+  }
+  public MaskingSessionUpdater firstParty(final String firstParty){
+    this.firstParty = firstParty;
+    return this;
+  }
+  public String secondParty() {
+    return this.secondParty;
+  }
+  public MaskingSessionUpdater secondParty(final String secondParty){
+    this.secondParty = secondParty;
+    return this;
+  }
   public int callTimeLimit() {
     return this.callTimeLimit;
   }
@@ -125,6 +140,13 @@ public class MaskingSessionUpdater extends VoiceUpdater<MaskingSessionUpdateResp
   }
   public MaskingSessionUpdater sessionExpiry(final int sessionExpiry) {
     this.sessionExpiry = sessionExpiry;
+    return this;
+  }
+  public boolean createSessionWithSingleParty() {
+    return this.createSessionWithSingleParty;
+  }
+  public MaskingSessionUpdater createSessionWithSingleParty(final boolean record){
+    this.createSessionWithSingleParty = createSessionWithSingleParty;
     return this;
   }
 }